actually move notificationInit out of the pack
it needs to be loaded by the initNotifications action, moving to backend.js

and rename to `load`, because `notificationsInit` vs `initNotifications` is just confusing :)
