+++ app/javascript/notifications/backend.js
@@ -1,3 +1,5 @@
+import { API } from '../http_api';
+
 export const maxNotifications = 100;
 
 export function convert(resource) {
@@ -18,4 +20,29 @@ export function convert(resource) {
     data,
     timeStamp: details.created_at || moment(new Date()).utc().format(),
   };
+}
+
+export function load(useLimit) {
+  const promises = [];
+  const limitFragment = useLimit ? `&limit=${maxNotifications}` : '';
+
+  promises.push(API.get(`/api/notifications?expand=resources&attributes=details&sort_by=id&sort_order=desc${limitFragment}`)
+    .then((data) => {
+      const notifications = data.resources.map(convert);
+
+      return {
+        notifications,
+        subcount: data.subcount,
+      };
+    }));
+
+  if (useLimit) {
+    // get real subcount
+    promises.push(API.get('/api/notifications'));
+  }
+
+  return Promise.all(promises).then(([{ notifications, subcount }, meta]) => ({
+    notifications,
+    subcount: meta ? meta.subcount : subcount,
+  }));
 }
